+++ svelte.config.js
@@ -1,18 +1,26 @@
 import { mdsvex } from 'mdsvex';
 import adapter from '@sveltejs/adapter-node';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
+import { fileURLToPath } from 'url';
+import { dirname, join } from 'path';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	// Consult https://svelte.dev/docs/kit/integrations
-	// for more information about preprocessors
-	preprocess: [vitePreprocess(), mdsvex()],
+    // Consult https://svelte.dev/docs/kit/integrations
+    // for more information about preprocessors
+    preprocess: [vitePreprocess(), mdsvex()],
 
-	kit: {
-		adapter: adapter()
-	},
+    kit: {
+        adapter: adapter(),
+        alias: {
+            $utils: join(__dirname, 'src/utils')
+        }
+    },
 
-	extensions: ['.svelte', '.svx']
+    extensions: ['.svelte', '.svx']
 };
 
 export default config;
