+++ vite.config.ts
@@ -2,9 +2,19 @@ import tailwindcss from '@tailwindcss/vite';
 import { svelteTesting } from '@testing-library/svelte/vite';
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
+import { fileURLToPath } from 'url';
+import { dirname, resolve } from 'path';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 export default defineConfig({
-	plugins: [sveltekit(), tailwindcss()],
+    plugins: [sveltekit(), tailwindcss()],
+    resolve: {
+        alias: {
+            $utils: resolve(__dirname, './src/utils')
+        }
+    },
 
 	test: {
 		workspace: [
